Aes (#53)

+++ tests/seismic/sol/AESEncrypt.sol
@@ -8,7 +8,7 @@ contract AESENCRYPT {
     /// @param plaintext The bytes to encrypt.
     /// @param nonce the u64 nonce for encryption, encoded as a big-endian bytes32.
     /// returns the encrypted bytes.
-    function AESEncrypt(bytes32 aes_key, uint256 nonce, bytes memory plaintext) public view returns (bytes memory) {
+    function AESEncrypt(bytes32 aes_key, uint96 nonce, bytes memory plaintext) public view returns (bytes memory) {
         // Address of the precompiled contract
         address AESEncryptAddr = address(0x67);
 
@@ -26,7 +26,7 @@ contract AESENCRYPT {
     }
 
     function testAESEncrypt() public view returns (bytes memory result) {
-        uint256 nonce = 17;
+        uint96 nonce = 17;
         bytes32 aes_key = hex"00112233445566778899aabbccddeeff00112233445566778899aabbccddeeff";
         bytes memory plaintext = hex"0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f";
         result = AESEncrypt(aes_key, nonce, plaintext);
@@ -37,4 +37,4 @@ contract AESENCRYPT {
 // EVMVersion: >=mercury
 // ====
 // ----
-// testAESEncrypt() -> hex"0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000004ff577b2b34b7dbafad6647accfaa9194d7a39c839e618fdbe9fc304691385c6fdcb1a8bf1c84560871726c31334884d85b463b0d9930c50370b9cdcc492dfcfb232dd38f0b0beb1c75e6f5c07e3a9ad0000000000000000000000000000000000"
+// testAESEncrypt() -> hex"0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000004f9915a626bc11d2f0bc37db72832936b6518c6b2e24a22467ba4259e03c1486f3e9110b69eb498f9213f11796b64b84eaabc145aaf209469a4d15698df70f0e8e98695496179a239d701803cfc922bd0000000000000000000000000000000000"
